🚚 글로벌 타입으로 통합
